requirements.txt
amqp==5.0.3
appdirs==1.4.4
asgiref==3.3.1
billiard==3.6.3.0
CacheControl==0.12.6
cachetools==4.2.1
celery==5.0.5
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
contextlib2==0.6.0.post1
coreapi==2.3.3
coreschema==0.0.4
coverage==5.3.1
cryptography==3.3.1
defusedxml==0.7.0rc1
distlib==0.3.1
distro==1.5.0
Django==3.1.5
django-cors-headers==3.5.0
django-stubs==1.7.0
django-templated-mail==1.1.1
django-unused-media==0.2.2
djangorestframework==3.12.2
djangorestframework-jsonapi==4.0.0
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
google-api-core==1.24.1
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
html5lib==1.0.1
httplib2==0.18.1
idna==2.8
inflection==0.5.1
ipaddr==2.2.0
itypes==1.2.0
Jinja2==2.11.2
kombu==5.0.2
lockfile==0.12.2
MarkupSafe==1.1.1
msgpack==0.6.2
mypy==0.790
mypy-extensions==0.4.3
numpy==1.19.4
oauthlib==3.1.0
packaging==20.3
pandas==1.1.5
pep517==0.8.2
Pillow==8.0.1
progress==1.5
prompt-toolkit==3.0.8
protobuf==3.14.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.6
python-dateutil==2.8.1
python3-openid==3.2.0
pytoml==0.1.21
pytz==2020.5
redis==3.5.3
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.6
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
spotipy==2.16.1
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.8
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
youtube-data-api==0.0.20